16.2.2-3

 Conflicts:
	cave/com.raytheon.viz.awipstools/src/com/raytheon/viz/awipstools/common/stormtrack/StormTrackDisplay.java
	cave/com.raytheon.viz.texteditor/src/com/raytheon/viz/texteditor/dialogs/TextEditorDialog.java
	cave/com.raytheon.viz.warngen/src/com/raytheon/viz/warngen/template/TemplateRunner.java
	edexOsgi/com.raytheon.edex.plugin.grib/src/com/raytheon/edex/plugin/grib/spatial/GribSpatialCache.java


Former-commit-id: 55bc369bddace873c371d0b5b51c383b33c6ca3d

+++ cave/com.raytheon.viz.awipstools/src/com/raytheon/viz/awipstools/ui/dialog/PutHomeCursorDialog.java
@@ -84,6 +84,7 @@ import com.vividsolutions.jts.io.WKTWriter;
  *                                    for the city/state issues.
  * Nov 11, 2014  3401     rferrel     Add Enter key events.
  * Jan 15, 2015  5054     randerso    Remove unnecessary new Shell
+ * Jan 16, 2016  DR 11474 A. Rickert  Parsing Lat/Lon with parseDouble for better accuracy
  * 
  * </pre>
  * 
@@ -471,7 +472,7 @@ public class PutHomeCursorDialog extends CaveSWTDialog implements
         } else if (latLonRadio.getSelection()) {
             Coordinate c = new Coordinate();
             try {
-                c.x = Float.parseFloat(lonTextField.getText());
+                c.x = Double.parseDouble(lonTextField.getText());
             } catch (NumberFormatException nfe) {
                 MessageDialog
                         .openError(shell, "Put Home Cursor Error",
@@ -480,7 +481,7 @@ public class PutHomeCursorDialog extends CaveSWTDialog implements
                 return;
             }
             try {
-                c.y = Float.parseFloat(latTextField.getText());
+                c.y = Double.parseDouble(latTextField.getText());
                 if ((c.y < -180) || c.y > 180) {
                     MessageDialog
                             .openError(shell, "Put Home Cursor Error",
